+++ InterferometricSignalsProcessing/Filters/ParticleFilterIS1D.h
@@ -8,19 +8,19 @@
 class ParticleFilterIS1D
 {
 public:
-	ParticleFilterIS1D(Eigen::Vector4d state_, Eigen::Matrix4d R_, Eigen::Matrix4d Rw_, double Rn_);
+	ParticleFilterIS1D(Eigen::Vector4d state_, Eigen::Matrix4d R_, Eigen::Matrix4d Rw_, float Rn_);
 	~ParticleFilterIS1D();
 
 	Eigen::Vector4d getState();
-	void estimate(double obs);
+	void estimate(float obs);
 private:
-	std::vector<double> state;
-	std::vector<double> means;
-	std::vector<double> vars;
+	std::vector<float> state;
+	std::vector<float> means;
+	std::vector<float> vars;
 
-	double Rn;
+	float Rn;
 
-	double h(Eigen::Vector4d state);
+	float h(Eigen::Vector4d state);
 	Eigen::Vector4d f(Eigen::Vector4d state);
 	Eigen::Matrix4d Ft(Eigen::Vector4d state);
 	Eigen::RowVector4d Ht(Eigen::Vector4d state);
